% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{as_kv}
\alias{as_kv}
\alias{as_kv.estimation_control}
\alias{as_kv.factor_model}
\alias{as_kv.model_component}
\alias{as_kv.model_system}
\title{Convert object to key-value format (internal)}
\usage{
as_kv(x, ...)

\method{as_kv}{estimation_control}(x, ...)

\method{as_kv}{factor_model}(x, ...)

\method{as_kv}{model_component}(x, ...)

\method{as_kv}{model_system}(x, ...)
}
\arguments{
\item{x}{Object to convert}

\item{...}{Additional arguments}
}
\description{
Convert object to key-value format (internal)
}
\keyword{internal}
