% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_estimation_control.R
\name{define_estimation_control}
\alias{define_estimation_control}
\title{Define estimation control settings}
\usage{
define_estimation_control(n_quad_points = 16, num_cores = 1)
}
\arguments{
\item{n_quad_points}{Integer. Number of Gauss-Hermite quadrature points for numerical integration (default = 16)}

\item{num_cores}{Integer. Number of processes to use for parallel estimation (default = 1)}
}
\value{
An object of class estimation_control containing control settings
}
\description{
Define estimation control settings
}
