% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disable_adaptive_quadrature_cpp}
\alias{disable_adaptive_quadrature_cpp}
\title{Disable adaptive quadrature}
\usage{
disable_adaptive_quadrature_cpp(fm_ptr)
}
\arguments{
\item{fm_ptr}{External pointer to FactorModel object}
}
\description{
Reverts to standard (non-adaptive) quadrature integration.
}
