% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_likelihood_cpp}
\alias{evaluate_likelihood_cpp}
\title{Evaluate log-likelihood for given parameters}
\usage{
evaluate_likelihood_cpp(
  fm_ptr,
  params,
  compute_gradient = FALSE,
  compute_hessian = FALSE
)
}
\arguments{
\item{fm_ptr}{External pointer to FactorModel object}

\item{params}{Vector of parameters}

\item{compute_gradient}{Whether to compute gradient (default FALSE)}

\item{compute_hessian}{Whether to compute Hessian (default FALSE)}
}
\value{
List with log-likelihood, gradient (if requested), and Hessian (if requested)
}
\description{
Evaluate log-likelihood for given parameters
}
