% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model_component.R
\name{get_factor}
\alias{get_factor}
\alias{get_factor.model_component}
\title{Get factor from component (internal)}
\usage{
get_factor(x, ...)

\method{get_factor}{model_component}(x, ...)
}
\arguments{
\item{x}{Object to extract factor from}

\item{...}{Additional arguments}
}
\description{
Get factor from component (internal)
}
\keyword{internal}
