% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print.components_table}
\alias{print.components_table}
\title{Print method for components_table}
\usage{
\method{print}{components_table}(x, ...)
}
\arguments{
\item{x}{A components_table object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for components_table
}
