% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_dyn_model_component.R
\name{print.dyn_model_component}
\alias{print.dyn_model_component}
\title{Print method for dynamic model component}
\usage{
\method{print}{dyn_model_component}(x, ...)
}
\arguments{
\item{x}{A dyn_model_component object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for dynamic model component
}
