% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{write_model_config_csv}
\alias{write_model_config_csv}
\title{Write a single CSV with all configuration rows}
\usage{
write_model_config_csv(model_system, factor_model, estimation_control, file)
}
\arguments{
\item{model_system}{a model_system object}

\item{factor_model}{a factor_model object}

\item{estimation_control}{an estimation_control object}

\item{file}{path to CSV to write}
}
\description{
Write a single CSV with all configuration rows
}
